fix(tests): extract pool exhaustion tests to separate file
test_heartbeat.py exceeded 500-line CI limit. Extracted 5
pool exhaustion tests to test_pool_exhaustion.py to stay
within limits.
