+++ tests/vibe3/runtime/test_heartbeat.py
@@ -472,190 +472,3 @@ async def test_tick_loop_skip_cleanup_when_disabled(monkeypatch) -> None:
     assert (
         len(cleanup_calls) == 0
     ), f"Expected no cleanup calls when disabled, got: {cleanup_calls}"
-
-
-class MockDispatchCoordinator:
-    """Mock coordinator for pool exhaustion testing."""
-
-    def __init__(self, is_paused: bool = False) -> None:
-        self._is_paused = is_paused
-
-    def is_dispatch_paused(self) -> bool:
-        """Return mocked pause state."""
-        return self._is_paused
-
-
-@pytest.mark.asyncio
-async def test_pool_exhaustion_counter_increments(monkeypatch) -> None:
-    """Counter should increment when pool is exhausted."""
-    coordinator = MockDispatchCoordinator(is_paused=True)
-    server = HeartbeatServer(
-        OrchestraConfig(
-            polling_interval=1,
-            max_concurrent_flows=3,
-            auto_stop_on_exhaustion=True,
-            exhaustion_threshold_ticks=10,  # Set high so it won't stop early
-        ),
-        dispatch_coordinator=coordinator,
-    )
-    svc = _MockService()
-    server.register(svc)
-
-    events: list[str] = []
-    tick_count = 0
-
-    def _capture(domain: str, message: str, **kwargs) -> None:
-        events.append(f"{domain}:{message}")
-
-    async def _sleep_once(_seconds: float) -> None:
-        nonlocal tick_count
-        tick_count += 1
-        if tick_count >= 6:  # Run 6 sleeps to allow 5 complete ticks
-            server.stop()
-
-    monkeypatch.setattr("vibe3.runtime.heartbeat.append_orchestra_event", _capture)
-    monkeypatch.setattr(
-        "vibe3.runtime.pool_exhaustion.append_orchestra_event", _capture
-    )
-    monkeypatch.setattr("vibe3.runtime.heartbeat.asyncio.sleep", _sleep_once)
-    server._running = True
-
-    await server._tick_loop()
-
-    # Counter should have incremented to 5 (after 5 complete ticks)
-    assert server._exhausted_ticks == 5
-    # Should have exhaustion events
-    exhaustion_events = [e for e in events if "pool exhausted" in e]
-    assert len(exhaustion_events) >= 5
-
-
-@pytest.mark.asyncio
-async def test_pool_exhaustion_counter_resets(monkeypatch) -> None:
-    """Counter should reset when pool has candidates."""
-    coordinator = MockDispatchCoordinator(is_paused=False)
-    server = HeartbeatServer(
-        OrchestraConfig(
-            polling_interval=1,
-            max_concurrent_flows=3,
-            auto_stop_on_exhaustion=True,
-            exhaustion_threshold_ticks=5,
-        ),
-        dispatch_coordinator=coordinator,
-    )
-    # Simulate previous exhaustion
-    server._exhausted_ticks = 4
-
-    svc = _MockService()
-    server.register(svc)
-
-    events: list[str] = []
-    tick_count = 0
-
-    def _capture(domain: str, message: str, **kwargs) -> None:
-        events.append(f"{domain}:{message}")
-
-    async def _sleep_once(_seconds: float) -> None:
-        nonlocal tick_count
-        tick_count += 1
-        if tick_count >= 2:
-            server.stop()
-
-    monkeypatch.setattr("vibe3.runtime.heartbeat.append_orchestra_event", _capture)
-    monkeypatch.setattr("vibe3.runtime.heartbeat.asyncio.sleep", _sleep_once)
-    server._running = True
-
-    await server._tick_loop()
-
-    # Counter should have reset to 0
-    assert server._exhausted_ticks == 0
-
-
-@pytest.mark.asyncio
-async def test_pool_exhaustion_stops_at_threshold(monkeypatch) -> None:
-    """Server should stop when exhaustion threshold reached."""
-    coordinator = MockDispatchCoordinator(is_paused=True)
-    server = HeartbeatServer(
-        OrchestraConfig(
-            polling_interval=1,
-            max_concurrent_flows=3,
-            auto_stop_on_exhaustion=True,
-            exhaustion_threshold_ticks=3,
-        ),
-        dispatch_coordinator=coordinator,
-    )
-    svc = _MockService()
-    server.register(svc)
-
-    events: list[str] = []
-
-    def _capture(domain: str, message: str, **kwargs) -> None:
-        events.append(f"{domain}:{message}")
-
-    async def _no_wait(_seconds: float) -> None:
-        return None
-
-    monkeypatch.setattr("vibe3.runtime.heartbeat.append_orchestra_event", _capture)
-    monkeypatch.setattr(
-        "vibe3.runtime.pool_exhaustion.append_orchestra_event", _capture
-    )
-    monkeypatch.setattr("vibe3.runtime.heartbeat.asyncio.sleep", _no_wait)
-    server._running = True
-
-    await server._tick_loop()
-
-    # Server should have stopped
-    assert server.running is False
-    # Counter should equal threshold
-    assert server._exhausted_ticks == 3
-    # Should have stop event
-    assert any("stopping server" in e for e in events)
-
-
-@pytest.mark.asyncio
-async def test_pool_exhaustion_no_coordinator_returns_false() -> None:
-    """Should return False when coordinator is None (no crash)."""
-    from vibe3.runtime.pool_exhaustion import is_pool_exhausted
-
-    assert is_pool_exhausted(None) is False
-
-
-@pytest.mark.asyncio
-async def test_pool_exhaustion_disabled_does_not_check(monkeypatch) -> None:
-    """Should not check exhaustion when auto_stop_on_exhaustion=False."""
-    coordinator = MockDispatchCoordinator(is_paused=True)
-    server = HeartbeatServer(
-        OrchestraConfig(
-            polling_interval=1,
-            max_concurrent_flows=3,
-            auto_stop_on_exhaustion=False,
-            exhaustion_threshold_ticks=3,
-        ),
-        dispatch_coordinator=coordinator,
-    )
-    svc = _MockService()
-    server.register(svc)
-
-    events: list[str] = []
-    tick_count = 0
-
-    def _capture(domain: str, message: str, **kwargs) -> None:
-        events.append(f"{domain}:{message}")
-
-    async def _sleep_once(_seconds: float) -> None:
-        nonlocal tick_count
-        tick_count += 1
-        if tick_count >= 5:
-            server.stop()
-
-    monkeypatch.setattr("vibe3.runtime.heartbeat.append_orchestra_event", _capture)
-    monkeypatch.setattr("vibe3.runtime.heartbeat.asyncio.sleep", _sleep_once)
-    server._running = True
-
-    await server._tick_loop()
-
-    # Should NOT have incremented counter
-    assert server._exhausted_ticks == 0
-    # Should NOT have exhaustion events
-    assert not any("pool exhausted" in e for e in events)
-    # Server should still be running until manual stop
-    assert tick_count == 5
